requirements.txt
pandas>=2.0
pyarrow>=14.0
great-expectations>=0.18
boto3>=1.34
pytest>=8.0
moto>=5.0
pyspark>=3.5
pyyaml>=6.0
